@@ -2,14 +2,14 @@
 
 namespace Stayallive\TLSA\Test;
 
-use PHPUnit\Framework\TestCase;
 use Stayallive\TLSA\Builder;
+use PHPUnit\Framework\TestCase;
 use Stayallive\TLSA\Exceptions\InvalidArgument;
 
 class BuilderTest extends TestCase
 {
     /** @test */
-    public function it_throws_an_exception_if_an_empty_string_is_passed(): void
+    public function itThrowsAnExceptionIfAnEmptyStringIsPassed(): void
     {
         $this->expectException(InvalidArgument::class);
         $this->expectExceptionMessage(InvalidArgument::domainIsMissing()->getMessage());
@@ -18,7 +18,7 @@ class BuilderTest extends TestCase
     }
 
     /** @test */
-    public function it_defaults_to_https_port(): void
+    public function itDefaultsToHttpsPort(): void
     {
         $builder = new Builder('alexbouma.me');
         $builder->forCertificate($this->get_certificate_stub());
@@ -27,7 +27,7 @@ class BuilderTest extends TestCase
     }
 
     /** @test */
-    public function it_detects_port_in_domain(): void
+    public function itDetectsPortInDomain(): void
     {
         $builder = new Builder('alexbouma.me:25');
         $builder->forCertificate($this->get_certificate_stub());
@@ -36,7 +36,7 @@ class BuilderTest extends TestCase
     }
 
     /** @test */
-    public function it_changes_protocol(): void
+    public function itChangesProtocol(): void
     {
         $builder = new Builder('alexbouma.me:25', 'udp');
         $builder->forCertificate($this->get_certificate_stub());
@@ -45,7 +45,7 @@ class BuilderTest extends TestCase
     }
 
     /** @test */
-    public function it_throws_an_exception_if_no_certificate_data_is_passed(): void
+    public function itThrowsAnExceptionIfNoCertificateDataIsPassed(): void
     {
         $this->expectException(InvalidArgument::class);
         $this->expectExceptionMessage(InvalidArgument::missingData()->getMessage());
@@ -54,7 +54,7 @@ class BuilderTest extends TestCase
     }
 
     /** @test */
-    public function it_throws_an_exception_if_invalid_certifiate_is_passed(): void
+    public function itThrowsAnExceptionIfInvalidCertifiateIsPassed(): void
     {
         $this->expectException(InvalidArgument::class);
         $this->expectExceptionMessage(InvalidArgument::invalidCertificate()->getMessage());
@@ -63,7 +63,7 @@ class BuilderTest extends TestCase
     }
 
     /** @test */
-    public function it_throws_an_exception_if_invalid_selector_is_used(): void
+    public function itThrowsAnExceptionIfInvalidSelectorIsUsed(): void
     {
         $this->expectException(InvalidArgument::class);
         $this->expectExceptionMessage(InvalidArgument::invalidSelectorForPublicKey()->getMessage());
@@ -74,7 +74,7 @@ class BuilderTest extends TestCase
     }
 
     /** @test */
-    public function it_throws_an_exception_if_invalid_matching_type_is_used(): void
+    public function itThrowsAnExceptionIfInvalidMatchingTypeIsUsed(): void
     {
         $this->expectException(InvalidArgument::class);
         $this->expectExceptionMessage(InvalidArgument::invalidMatchingType()->getMessage());
@@ -86,18 +86,19 @@ class BuilderTest extends TestCase
     }
 
     /** @test */
-    public function it_builds_a_valid_record_with_certificate_3_0_0(): void
+    public function itBuildsAValidRecordWithCertificate300(): void
     {
         $builder = (new Builder('alexbouma.me'))
             ->selector(Builder::SELECTOR_CERTIFICATE)
             ->matchingType(Builder::MATCHING_TYPE_FULL)
             ->forCertificate($this->get_certificate_stub());
 
-        $this->assertEquals('_443._tcp.alexbouma.me. IN TLSA 3 0 0 308206353082051da0030201020212033db6b3a14f9afbb87d0752866bcbf10741300d06092a864886f70d01010b0500304a310b300906035504061302555331163014060355040a130d4c6574277320456e6372797074312330210603550403131a4c6574277320456e637279707420417574686f72697479205833301e170d3138303831383139323031325a170d3138313131363139323031325a3017311530130603550403130c616c6578626f756d612e6d6530820122300d06092a864886f70d01010105000382010f003082010a0282010100caaf466cd7ef4387ffddb7644ccff8d8453d5051e77bd5b7653302908e29b17548ad2eaa500bf8c9a7eb60fd2eab9e14f17e572e2c42ef15844fa39e91a62bcabf2b2e76103dee10385657c696bf1aa90b31d56f3026e24341b0ee6b897f752f21f022fb85890483fd793181b5891ac50d4fda45ba9c9732621c0bbb58a9f2da48f371cd53aab05202b4a425ffd180222e9a136b6b5849389a7fcc4a550492965b2c1287ebf57a6ce9d16d532cba65c1c4b5afdec676165b35acc9066efa32f29146de5b7811f19c9474d4364fed41f4d9ab4c8056649f4938c2375b9a32d70dfbe3d9153f580c64a881d8e5e3190e998e5cd8accd9aa167d1c543e4cecd08ad0203010001a382034630820342300e0603551d0f0101ff0404030205a0301d0603551d250416301406082b0601050507030106082b06010505070302300c0603551d130101ff04023000301d0603551d0e04160414a95815fc787e637ab536a2f3a217a84e0d1cb620301f0603551d23041830168014a84a6a63047dddbae6d139b7a64565eff3a8eca1306f06082b0601050507010104633061302e06082b060105050730018622687474703a2f2f6f6373702e696e742d78332e6c657473656e63727970742e6f7267302f06082b060105050730028623687474703a2f2f636572742e696e742d78332e6c657473656e63727970742e6f72672f30490603551d1104423040820c616c6578626f756d612e6d65820c616c6578626f756d612e6e6c82107777772e616c6578626f756d612e6d6582107777772e616c6578626f756d612e6e6c3081fe0603551d200481f63081f33008060667810c0102013081e6060b2b0601040182df130101013081d6302606082b06010505070201161a687474703a2f2f6370732e6c657473656e63727970742e6f72673081ab06082b0601050507020230819e0c819b54686973204365727469666963617465206d6179206f6e6c792062652072656c6965642075706f6e2062792052656c79696e67205061727469657320616e64206f6e6c7920696e206163636f7264616e636520776974682074686520436572746966696361746520506f6c69637920666f756e642061742068747470733a2f2f6c657473656e63727970742e6f72672f7265706f7369746f72792f30820104060a2b06010401d6790204020481f50481f200f0007700c1164ae0a772d2d4392dc80ac10770d4f0c49bde991a4840c1fa075164f63360000001654eb1fdeb0000040300483046022100cba0b03379cd00ebff907bc083482809842ad510e9b0cabb75a41b31fff087150221009ec7492aa00846d750785ea7a6f7089be8c74f9bc009e357dba8a6c3c5243bf2007500293c519654c83965baaa50fc5807d4b76fbf587a2972dca4c30cf4e54547f478000001654eb1fe870000040300463044022022c60c0b3948107e3ca3ea7712c3f848fb63ef7ac9b33eb6df1b61398f7a944d02201c4b3a25173b21fd0c8dcea339557737c233c58f54b0f5574796b576230a8b0f300d06092a864886f70d01010b050003820101007517c331cc7375dcd86ae20d36684ce96a25a8ac1d58a566f17f3ea11dcf5c90b04b27ccabba9ed83bcbc6e46194e1314d59822898dbf288bf1596980c1fa98b6943c8e97516f4efc6ccab0eac7d33c0c2f545aa2bcc73dcb78cc0864e950190eeafab9629831d02d616eaa4f292aa17857abf607cad40995622d40b6329727d5c1e788f1c6372273056bcd881c1fad2169ccf6d8ca1e49fc839272ecd206d3c212b816039c1d17834c4a8d9ccc54b2daddac22f5bbf0c4f528176ac7b0d2a06fb54ab9ddf7331360910cc0e52bfe887da307c3caf1d0f8ff65c30a09815699599905399af2f1300f517b29c3ddf798745404b715552e681a539d18843a7f5ce', $builder->getRecord());
+        $this->assertEquals('_443._tcp.alexbouma.me. IN TLSA 3 0 0 308206353082051da0030201020212033db6b3a14f9afbb87d0752866bcbf10741300d06092a864886f70d01010b0500304a310b300906035504061302555331163014060355040a130d4c6574277320456e6372797074312330210603550403131a4c6574277320456e637279707420417574686f72697479205833301e170d3138303831383139323031325a170d3138313131363139323031325a3017311530130603550403130c616c6578626f756d612e6d6530820122300d06092a864886f70d01010105000382010f003082010a0282010100caaf466cd7ef4387ffddb7644ccff8d8453d5051e77bd5b7653302908e29b17548ad2eaa500bf8c9a7eb60fd2eab9e14f17e572e2c42ef15844fa39e91a62bcabf2b2e76103dee10385657c696bf1aa90b31d56f3026e24341b0ee6b897f752f21f022fb85890483fd793181b5891ac50d4fda45ba9c9732621c0bbb58a9f2da48f371cd53aab05202b4a425ffd180222e9a136b6b5849389a7fcc4a550492965b2c1287ebf57a6ce9d16d532cba65c1c4b5afdec676165b35acc9066efa32f29146de5b7811f19c9474d4364fed41f4d9ab4c8056649f4938c2375b9a32d70dfbe3d9153f580c64a881d8e5e3190e998e5cd8accd9aa167d1c543e4cecd08ad0203010001a382034630820342300e0603551d0f0101ff0404030205a0301d0603551d250416301406082b0601050507030106082b06010505070302300c0603551d130101ff04023000301d0603551d0e04160414a95815fc787e637ab536a2f3a217a84e0d1cb620301f0603551d23041830168014a84a6a63047dddbae6d139b7a64565eff3a8eca1306f06082b0601050507010104633061302e06082b060105050730018622687474703a2f2f6f6373702e696e742d78332e6c657473656e63727970742e6f7267302f06082b060105050730028623687474703a2f2f636572742e696e742d78332e6c657473656e63727970742e6f72672f30490603551d1104423040820c616c6578626f756d612e6d65820c616c6578626f756d612e6e6c82107777772e616c6578626f756d612e6d6582107777772e616c6578626f756d612e6e6c3081fe0603551d200481f63081f33008060667810c0102013081e6060b2b0601040182df130101013081d6302606082b06010505070201161a687474703a2f2f6370732e6c657473656e63727970742e6f72673081ab06082b0601050507020230819e0c819b54686973204365727469666963617465206d6179206f6e6c792062652072656c6965642075706f6e2062792052656c79696e67205061727469657320616e64206f6e6c7920696e206163636f7264616e636520776974682074686520436572746966696361746520506f6c69637920666f756e642061742068747470733a2f2f6c657473656e63727970742e6f72672f7265706f7369746f72792f30820104060a2b06010401d6790204020481f50481f200f0007700c1164ae0a772d2d4392dc80ac10770d4f0c49bde991a4840c1fa075164f63360000001654eb1fdeb0000040300483046022100cba0b03379cd00ebff907bc083482809842ad510e9b0cabb75a41b31fff087150221009ec7492aa00846d750785ea7a6f7089be8c74f9bc009e357dba8a6c3c5243bf2007500293c519654c83965baaa50fc5807d4b76fbf587a2972dca4c30cf4e54547f478000001654eb1fe870000040300463044022022c60c0b3948107e3ca3ea7712c3f848fb63ef7ac9b33eb6df1b61398f7a944d02201c4b3a25173b21fd0c8dcea339557737c233c58f54b0f5574796b576230a8b0f300d06092a864886f70d01010b050003820101007517c331cc7375dcd86ae20d36684ce96a25a8ac1d58a566f17f3ea11dcf5c90b04b27ccabba9ed83bcbc6e46194e1314d59822898dbf288bf1596980c1fa98b6943c8e97516f4efc6ccab0eac7d33c0c2f545aa2bcc73dcb78cc0864e950190eeafab9629831d02d616eaa4f292aa17857abf607cad40995622d40b6329727d5c1e788f1c6372273056bcd881c1fad2169ccf6d8ca1e49fc839272ecd206d3c212b816039c1d17834c4a8d9ccc54b2daddac22f5bbf0c4f528176ac7b0d2a06fb54ab9ddf7331360910cc0e52bfe887da307c3caf1d0f8ff65c30a09815699599905399af2f1300f517b29c3ddf798745404b715552e681a539d18843a7f5ce',
+            $builder->getRecord());
     }
 
     /** @test */
-    public function it_builds_a_valid_record_with_certificate_3_0_1(): void
+    public function itBuildsAValidRecordWithCertificate301(): void
     {
         $builder = (new Builder('alexbouma.me'))
             ->selector(Builder::SELECTOR_CERTIFICATE)
@@ -108,7 +109,7 @@ class BuilderTest extends TestCase
     }
 
     /** @test */
-    public function it_builds_a_valid_record_with_certificate_3_0_2(): void
+    public function itBuildsAValidRecordWithCertificate302(): void
     {
         $builder = (new Builder('alexbouma.me'))
             ->selector(Builder::SELECTOR_CERTIFICATE)
@@ -119,18 +120,19 @@ class BuilderTest extends TestCase
     }
 
     /** @test */
-    public function it_builds_a_valid_record_with_certificate_3_1_0(): void
+    public function itBuildsAValidRecordWithCertificate310(): void
     {
         $builder = (new Builder('alexbouma.me'))
             ->selector(Builder::SELECTOR_PUBLIC_KEY)
             ->matchingType(Builder::MATCHING_TYPE_FULL)
             ->forCertificate($this->get_certificate_stub());
 
-        $this->assertEquals('_443._tcp.alexbouma.me. IN TLSA 3 1 0 30820122300d06092a864886f70d01010105000382010f003082010a0282010100caaf466cd7ef4387ffddb7644ccff8d8453d5051e77bd5b7653302908e29b17548ad2eaa500bf8c9a7eb60fd2eab9e14f17e572e2c42ef15844fa39e91a62bcabf2b2e76103dee10385657c696bf1aa90b31d56f3026e24341b0ee6b897f752f21f022fb85890483fd793181b5891ac50d4fda45ba9c9732621c0bbb58a9f2da48f371cd53aab05202b4a425ffd180222e9a136b6b5849389a7fcc4a550492965b2c1287ebf57a6ce9d16d532cba65c1c4b5afdec676165b35acc9066efa32f29146de5b7811f19c9474d4364fed41f4d9ab4c8056649f4938c2375b9a32d70dfbe3d9153f580c64a881d8e5e3190e998e5cd8accd9aa167d1c543e4cecd08ad0203010001', $builder->getRecord());
+        $this->assertEquals('_443._tcp.alexbouma.me. IN TLSA 3 1 0 30820122300d06092a864886f70d01010105000382010f003082010a0282010100caaf466cd7ef4387ffddb7644ccff8d8453d5051e77bd5b7653302908e29b17548ad2eaa500bf8c9a7eb60fd2eab9e14f17e572e2c42ef15844fa39e91a62bcabf2b2e76103dee10385657c696bf1aa90b31d56f3026e24341b0ee6b897f752f21f022fb85890483fd793181b5891ac50d4fda45ba9c9732621c0bbb58a9f2da48f371cd53aab05202b4a425ffd180222e9a136b6b5849389a7fcc4a550492965b2c1287ebf57a6ce9d16d532cba65c1c4b5afdec676165b35acc9066efa32f29146de5b7811f19c9474d4364fed41f4d9ab4c8056649f4938c2375b9a32d70dfbe3d9153f580c64a881d8e5e3190e998e5cd8accd9aa167d1c543e4cecd08ad0203010001',
+            $builder->getRecord());
     }
 
     /** @test */
-    public function it_builds_a_valid_record_with_certificate_3_1_1(): void
+    public function itBuildsAValidRecordWithCertificate311(): void
     {
         $builder = (new Builder('alexbouma.me'))
             ->selector(Builder::SELECTOR_PUBLIC_KEY)
@@ -141,7 +143,7 @@ class BuilderTest extends TestCase
     }
 
     /** @test */
-    public function it_builds_a_valid_record_with_certificate_3_1_2(): void
+    public function itBuildsAValidRecordWithCertificate312(): void
     {
         $builder = (new Builder('alexbouma.me'))
             ->selector(Builder::SELECTOR_PUBLIC_KEY)
@@ -152,7 +154,7 @@ class BuilderTest extends TestCase
     }
 
     /** @test */
-    public function it_builds_a_valid_record_with_certificate_0_1_1(): void
+    public function itBuildsAValidRecordWithCertificate011(): void
     {
         $builder = (new Builder('alexbouma.me'))
             ->certificateUsage(Builder::CERTIFICATE_USAGE_CA)
@@ -164,7 +166,7 @@ class BuilderTest extends TestCase
     }
 
     /** @test */
-    public function it_builds_a_valid_record_with_certificate_1_1_1(): void
+    public function itBuildsAValidRecordWithCertificate111(): void
     {
         $builder = (new Builder('alexbouma.me'))
             ->certificateUsage(Builder::CERTIFICATE_USAGE_SERVICE_CERTIFICATE)
@@ -176,7 +178,7 @@ class BuilderTest extends TestCase
     }
 
     /** @test */
-    public function it_builds_a_valid_record_with_certificate_2_1_1(): void
+    public function itBuildsAValidRecordWithCertificate211(): void
     {
         $builder = (new Builder('alexbouma.me'))
             ->certificateUsage(Builder::CERTIFICATE_USAGE_TRUST_ANCHOR_ASSERTION)
@@ -189,11 +191,11 @@ class BuilderTest extends TestCase
 
     private function get_certificate_stub(): string
     {
-        return file_get_contents(__DIR__.'/stubs/alexbouma_me.crt');
+        return file_get_contents(__DIR__ . '/stubs/alexbouma_me.crt');
     }
 
     private function get_public_key_stub(): string
     {
-        return file_get_contents(__DIR__.'/stubs/alexbouma_me.pub');
+        return file_get_contents(__DIR__ . '/stubs/alexbouma_me.pub');
     }
 }
